Delete designSystem folder and move all stories components, Refactoring components types, structure
