+++ client/app/stories/components/index.ts
@@ -1,3 +1,4 @@
-export * from './designSystem/BorderRadiuses'
-
+export * from './BorderRadiuses'
+export * from './BoxShadows'
+export * from './DesignSystemPlate'
 export * from './Example'
